server/requirements.txt
flask
python-telegram-bot==13.14
gradio_client
python-dotenv